Refactor workbar into modular architecture
