+++ src/modules/hyprland/workbar/model.cpp
@@ -1,0 +1,1 @@
+#include "modules/hyprland/workbar/model.hpp"
